black styling

+++ utils/data_preprocessor.py
@@ -6,7 +6,7 @@ import tqdm
 
 def cmp_to_key(mycmp):
     """
-        Convert a cmp= function into a key= function
+    Convert a cmp= function into a key= function
     """
 
     class K(object):
@@ -55,15 +55,15 @@ def label_dataset(path_to_files, files):
     evnts = {}
     cur = 0
     for i, f in tqdm.tqdm(enumerate(files)):
-        print('File: {}'.format(i))
-        df = pd.read_csv(path_to_files + '/' + f)
-        for i in range(len(df['event'])):
-            if df['event'].iloc[i] not in evnts:
-                evnts[df['event'].iloc[i]] = cur
+        print("File: {}".format(i))
+        df = pd.read_csv(path_to_files + "/" + f)
+        for i in range(len(df["event"])):
+            if df["event"].iloc[i] not in evnts:
+                evnts[df["event"].iloc[i]] = cur
                 cur += 1
-            df['event'].iloc[i] = evnts[df['event'].iloc[i]]
-        df['event'] = df['event'].astype(int)
-        df.to_csv(path_to_files + '/' + f)
+            df["event"].iloc[i] = evnts[df["event"].iloc[i]]
+        df["event"] = df["event"].astype(int)
+        df.to_csv(path_to_files + "/" + f)
 
 
 evnts = {}
@@ -72,23 +72,23 @@ cur = 0
 
 def get_partition(df, num_of_steps, num_of_classes, end_time=None):
     """
-        Transforms dataset into partition
+    Transforms dataset into partition
 
-        inputs:
-                df - pandas.DataFrame, columns - time and event
-                num_of_steps - int, number of steps in partition
-                num_of_classes - int, number of event types
-                end_time - float, end time or None
+    inputs:
+            df - pandas.DataFrame, columns - time and event
+            num_of_steps - int, number of steps in partition
+            num_of_classes - int, number of event types
+            end_time - float, end time or None
 
-        outputs:
-                partition - torch.Tensor, size = (num_of_steps, num_of_classes + 1)
+    outputs:
+            partition - torch.Tensor, size = (num_of_steps, num_of_classes + 1)
     """
-    df = df.loc[:, ['time', 'event']].copy()
-    df = df.sort_values(by=['time'])
-    df = df.reset_index().loc[:, ['time', 'event']].copy()
+    df = df.loc[:, ["time", "event"]].copy()
+    df = df.sort_values(by=["time"])
+    df = df.reset_index().loc[:, ["time", "event"]].copy()
     # setting end time if None
     if end_time is None:
-        end_time = df['time'][len(df['time']) - 1]
+        end_time = df["time"][len(df["time"]) - 1]
 
     # preparing output template
     res = torch.zeros(num_of_steps, num_of_classes + 1)
@@ -98,26 +98,26 @@ def get_partition(df, num_of_steps, num_of_classes, end_time=None):
     res[:, 0] = end_time / num_of_steps
 
     # converting time to timestamps
-    df['time'] = (df['time'] / dt).astype(int)
-    mask = (df['time'] == num_of_steps)
-    df.loc[mask, 'time'] -= 1
+    df["time"] = (df["time"] / dt).astype(int)
+    mask = df["time"] == num_of_steps
+    df.loc[mask, "time"] -= 1
 
     # counting points
     df = df.reset_index()
-    df = df.groupby(['time', 'event']).count()
+    df = df.groupby(["time", "event"]).count()
     df = df.reset_index()
-    df.columns = ['time', 'event', 'num']
+    df.columns = ["time", "event", "num"]
     try:
-        df['event'] = df['event'].astype(int)
+        df["event"] = df["event"].astype(int)
     except:
         global evnts
         global cur
-        for i in range(len(df['event'])):
-            if df['event'].iloc[i] not in evnts:
-                evnts[df['event'].iloc[i]] = cur
+        for i in range(len(df["event"])):
+            if df["event"].iloc[i] not in evnts:
+                evnts[df["event"].iloc[i]] = cur
                 cur += 1
-            df['event'].iloc[i] = evnts[df['event'].iloc[i]]
-        df['event'] = df['event'].astype(int)
+            df["event"].iloc[i] = evnts[df["event"].iloc[i]]
+        df["event"] = df["event"].astype(int)
 
     # computing partition
     tmp = torch.Tensor(df.to_numpy()).long()
@@ -128,16 +128,16 @@ def get_partition(df, num_of_steps, num_of_classes, end_time=None):
 
 def get_dataset(path_to_files: str, n_classes: int, n_steps: int, n_files=None):
     """
-        Reads dataset
+    Reads dataset
 
-        inputs:
-                path_to_files - str, path to csv files with dataset
-                n_classes - int, number of event types
-                n_steps - int, number of steps in partitions
+    inputs:
+            path_to_files - str, path to csv files with dataset
+            n_classes - int, number of event types
+            n_steps - int, number of steps in partitions
 
-        outputs:
-                data - torch.Tensor, size = (N, n_steps, n_classes + 1), dataset
-                target - torch.Tensor, size = (N), true labels or None
+    outputs:
+            data - torch.Tensor, size = (N, n_steps, n_classes + 1), dataset
+            target - torch.Tensor, size = (N), true labels or None
     """
     # searching for files
     files = os.listdir(path_to_files)
@@ -145,13 +145,15 @@ def get_dataset(path_to_files: str, n_classes: int, n_steps: int, n_files=None):
     last_event_target = False
 
     # reading target
-    if 'clusters.csv' in files:
-        files.remove('clusters.csv')
-        target = torch.Tensor(pd.read_csv(path_to_files + '/clusters.csv')['cluster_id'])
+    if "clusters.csv" in files:
+        files.remove("clusters.csv")
+        target = torch.Tensor(
+            pd.read_csv(path_to_files + "/clusters.csv")["cluster_id"]
+        )
         if n_files is not None:
             target = target[:n_files]
-    if 'info.json' in files:
-        files.remove('info.json')
+    if "info.json" in files:
+        files.remove("info.json")
 
     # reading data
     files = sorted(files, key=cmp_to_key(compare))
@@ -159,8 +161,8 @@ def get_dataset(path_to_files: str, n_classes: int, n_steps: int, n_files=None):
         files = files[:n_files]
     data = torch.zeros(len(files), n_steps, n_classes + 1)
     for i, f in tqdm.tqdm(enumerate(files)):
-        print('File: {}'.format(f))
-        df = pd.read_csv(path_to_files + '/' + f)
+        print("File: {}".format(f))
+        df = pd.read_csv(path_to_files + "/" + f)
         data[i, :, :] = get_partition(df, n_steps, n_classes)
 
     return data, target
